+++ ext/gb/gb_math.c
@@ -1,2 +1,0 @@
-#define GB_MATH_IMPLEMENTATION
-#include "gb_math.h"