+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 - Remove `ext/aom` as a submodule. If libavif users want to build aom from ext/, they must enable `AVIF_BUILD_AOM` and supply their own local copy.
+- Move the handful of public domain gb_math functions used by colr.c and eliminate the dependence on the gb library
 
 ## [0.3.1] - 2019-07-22
 ### Changed